+++ src/assets/scripts/aircraft/AircraftInstanceModel.js
@@ -6,9 +6,9 @@ import _has from 'lodash/has';
 import _isNaN from 'lodash/isNaN';
 import _isString from 'lodash/isString';
 import _map from 'lodash/map';
-import AircraftFlightManagementSystem from './AircraftFlightManagementSystem';
+import AircraftFlightManagementSystem from './FlightManagementSystem/AircraftFlightManagementSystem';
 import AircraftStripView from './AircraftStripView';
-import Waypoint from './Waypoint';
+import Waypoint from './FlightManagementSystem/Waypoint';
 import { speech_say } from '../speech';
 import { tau, radians_normalize, angle_offset } from '../math/circle';
 import { round, abs, sin, cos, extrapolate_range_clamp, clamp } from '../math/core';
