+++ src/assets/scripts/aircraft/FlightManagementSystem/Leg.js
@@ -3,9 +3,9 @@ import _get from 'lodash/get';
 import _has from 'lodash/has';
 import _map from 'lodash/map';
 import Waypoint from './Waypoint';
-import RouteModel from '../airport/Route/RouteModel';
-import { FP_LEG_TYPE } from '../constants/aircraftConstants';
-import { LOG } from '../constants/logLevel';
+import RouteModel from '../../airport/Route/RouteModel';
+import { FP_LEG_TYPE } from '../../constants/aircraftConstants';
+import { LOG } from '../../constants/logLevel';
 
 /**
   * This class acts as a collection of Waypoint model objects.
